Add sync endpoint. Reorganize types
